+++ apps/client/src/app/pages/pricing/pricing-page.module.ts
@@ -1,6 +1,8 @@
 import { CommonModule } from '@angular/common';
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
+import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
+import { RouterModule } from '@angular/router';
 
 import { PricingPageRoutingModule } from './pricing-page-routing.module';
 import { PricingPageComponent } from './pricing-page.component';
@@ -8,7 +10,13 @@ import { PricingPageComponent } from './pricing-page.component';
 @NgModule({
   declarations: [PricingPageComponent],
   exports: [],
-  imports: [CommonModule, MatCardModule, PricingPageRoutingModule],
+  imports: [
+    CommonModule,
+    MatButtonModule,
+    MatCardModule,
+    PricingPageRoutingModule,
+    RouterModule
+  ],
   providers: [],
   schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })