+++ libs/common/src/lib/types/subscription.type.ts
@@ -1,0 +1,4 @@
+export enum SubscriptionType {
+  Basic = 'Basic',
+  Premium = 'Premium'
+}